String.prototype.length is not a method
Line 6 indicates that the string variable "str" has a method called length(). String.prototype, does not have a length() method, it has a property called length. 
Also that code block for ES5 string methods look quite unclear, I included a numbered list with a small description, after all it's to remind the reader and a hyperlink to the MDN docs to read more about it.